requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
huggingface-hub==0.19.4
openai==1.3.5
python-logging-config==0.1.0
